requirements.txt
streamlit
Pillow
firebase-admin
anthropic
openai
pygame
sounddevice
soundfile
scipy
numpy
nltk
google-generativeai
replicate
python-dotenv
pyttsx3
SpeechRecognition
pyaudio
